+++ src/Symfony/Component/RateLimiter/Policy/TokenBucketLimiter.php
@@ -21,8 +21,6 @@ use Symfony\Component\RateLimiter\Storage\StorageInterface;
 
 /**
  * @author Wouter de Jong <wouter@wouterj.nl>
- *
- * @experimental in 5.3
  */
 final class TokenBucketLimiter implements LimiterInterface
 {