+++ src/Symfony/Component/Translation/Bridge/Crowdin/CrowdinProvider.php
@@ -28,8 +28,6 @@ use Symfony\Contracts\HttpClient\ResponseInterface;
  *  * Filenames refer to Symfony's translation domains;
  *  * Identifiers refer to Symfony's translation keys;
  *  * Translations refer to Symfony's translated messages
- *
- * @experimental in 5.3
  */
 final class CrowdinProvider implements ProviderInterface
 {