+++ src/Symfony/Component/Translation/Bridge/Crowdin/CrowdinProviderFactory.php
@@ -22,8 +22,6 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
  * @author Andrii Bodnar <andrii.bodnar@crowdin.com>
- *
- * @experimental in 5.3
  */
 final class CrowdinProviderFactory extends AbstractProviderFactory
 {