+++ src/Symfony/Component/Translation/Bridge/Loco/LocoProvider.php
@@ -26,8 +26,6 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
  *  * Tags refers to Symfony's translation domains
  *  * Assets refers to Symfony's translation keys
  *  * Translations refers to Symfony's translated messages
- *
- * @experimental in 5.3
  */
 final class LocoProvider implements ProviderInterface
 {