+++ src/Symfony/Component/Translation/Bridge/Lokalise/LokaliseProviderFactory.php
@@ -20,8 +20,6 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
  * @author Mathieu Santostefano <msantostefano@protonmail.com>
- *
- * @experimental in 5.3
  */
 final class LokaliseProviderFactory extends AbstractProviderFactory
 {