+++ src/Symfony/Component/Translation/Provider/FilteringProvider.php
@@ -18,8 +18,6 @@ use Symfony\Component\Translation\TranslatorBagInterface;
  * Filters domains and locales between the Translator config values and those specific to each provider.
  *
  * @author Mathieu Santostefano <msantostefano@protonmail.com>
- *
- * @experimental in 5.3
  */
 class FilteringProvider implements ProviderInterface
 {