+++ src/Symfony/Component/Translation/CHANGELOG.md
@@ -4,8 +4,9 @@ CHANGELOG
 5.4
 ---
 
-* Add `github` format & autodetection to render errors as annotations when
-  running the XLIFF linter command in a Github Actions environment.
+ * Add `github` format & autodetection to render errors as annotations when
+   running the XLIFF linter command in a Github Actions environment.
+ * Translation providers are not experimental anymore
 
 5.3
 ---